adding option to provide custom ca-certs for the ssl connection
Implemented a --ca-certificate option to point to a file containing
ca-certs.  This change fixes an issue where ``curl`` is not happy with
the security when accessing the server.  It can work by using the ``-k``
option on ``curl`` to disable security, but that is not ideal for a
testing setup.

Signed-off-by: Mark Nicholson <mark.j.nicholson@oracle.com>